+++ src/mmio.c
@@ -1,0 +1,6 @@
+#include "mmio.h"
+
+void
+mmio_translation_unit_anchor(void)
+{
+}